lvm: Remove --select usage from reload pvs
Revert back to using "pvs [pv1, pv2, ...]" for reloading lvm cache.
Using --select option for lvm command requires a scan of all VGs
metadata before it can process only the relevant devices for the selected
entities. In vdsm 4.3 with lvm2-2.02 this induces the risk for transient
corruption in VG metadata for reload commands, for vdsm 4.4 with lvm2-2.03
this is not reproducible but there is a performance overhead for using
lvm --select commands. For sake of stability in released and undergoing
release versions we revert back to the old way.

This implies that reload with "pvs pv1 pv2" where pv1 is missing from
LVM will cause the entire command to fail, marking pv2 as Unreadable
and return those for getPV and getAllPVs methods, hence stale PV tests
that used to pass with pvs --select for reloads are modified.

Change-Id: I73e6b2ca55862d41ccb48b38c33003d4eacb4c3c
Bug-Url: https://bugzilla.redhat.com/1849595
Signed-off-by: Amit Bawer <abawer@redhat.com>