lvmfilter: Add function for finding disk devices of lvm devices
This will allow to query the lvmfilter module for the devices underlying
the /dev/mapper devices used for the local PVs we want to add to the
lvm filter.

We gather only disk devices for non-multipath devices as we do not want
to blacklist already configured multipath devices on the host, but
preventing it from claiming local devices which were not originally
managed by multipath.

The disk devices WWIDs will be added to multipath blacklist config,
so their names would remain consistent during reboots and upgrades.

Change-Id: Ia6f22819d0c3891a300ef51a84253de1955aa6fc
Bug-Url: https://bugzilla.redhat.com/1837864
Signed-off-by: Amit Bawer <abawer@redhat.com>